Adds subscribe user to many spaces
